QuestLogMaker_Proto
This is a prototype and is not ready for use/release. If anyone wants to
edit this go ahead but I doubt any of you will want to hahaha! This is
also a test to see if I can make a programming branch separate from the
main that I can work on!